change layout and separate main content into sub component
