+++ client/src/pages/Article/Articles.jsx
@@ -1,6 +1,8 @@
 import React, { memo } from "react";
+
 import people from "../../assets/images/people.jpg";
 import sports from "../../assets/images/sports.jpg";
+import Grid from "../../component/Layout/card/Grid";
 import {
   Bookmark,
   Share2,
@@ -125,8 +127,6 @@ function Article() {
                 </svg>
               </button>
             </div>
-
-            {/* Stats Badge */}
             <div className="hidden md:flex items-center space-x-1 text-sm text-gray-600 dark:text-gray-400">
               <span>12 Published</span>
               <span>•</span>
@@ -160,8 +160,6 @@ function Article() {
                   {["Business", "Sports", "Tech", "Health"][i % 4]}
                 </span>
               </div>
-
-              {/* Article content */}
               <div className="p-4">
                 <div className="flex items-center gap-2 text-xs text-gray-500 dark:text-gray-400 mb-2">
                   <div className="flex items-center gap-1">
@@ -184,8 +182,6 @@ function Article() {
                   The latest automotive innovations are making waves before they
                   even hit the showroom floor at this year's event.
                 </p>
-
-                {/* Stats and interactions */}
                 <div className="flex justify-between items-center pt-2 border-t border-gray-100 dark:border-gray-700">
                   <div className="flex gap-3 text-xs text-gray-500 dark:text-gray-400">
                     <span className="flex items-center gap-1">
@@ -209,8 +205,6 @@ function Article() {
             </Link>
           ))}
         </div>
-
-        {/* Load more button */}
         <div className="flex justify-center mt-10">
           <button className="px-6 py-2 bg-white dark:bg-gray-800 border border-gray-200 dark:border-gray-700 rounded-full text-sm font-medium text-gray-700 dark:text-gray-300 hover:bg-gray-100 dark:hover:bg-gray-700 transition-colors">
             Load More Articles
