+++ client/src/routes/routes.js
@@ -6,7 +6,7 @@ import Blogs from "../pages/Article/Article";
 import LiveReports from "../pages/VIdeo/VIdeos";
 import Saved from "../pages/saved/Saved";
 import StreamDetails from "../pages/VIdeo/StreamDetails";
-import Opinion from "../pages/Opinion";
+import Opinion from "../pages/opinion/Opinion";
 import SpecialDay from "../pages/SpecialDay";
 import History from "../pages/Histrory";
 import Profile from "../pages/profile/Profile";
